+++ tests/Provider/Time/FixedTimeProviderTest.php
@@ -24,17 +24,17 @@ class FixedTimeProviderTest extends TestCase
         $provider = new FixedTimeProvider($time);
 
         $this->assertSame('1458844556', $provider->getTime()->getSeconds()->toString());
-        $this->assertSame('200997', $provider->getTime()->getMicroSeconds()->toString());
+        $this->assertSame('200997', $provider->getTime()->getMicroseconds()->toString());
 
         $provider->setSec(1050804050);
 
         $this->assertSame('1050804050', $provider->getTime()->getSeconds()->toString());
-        $this->assertSame('200997', $provider->getTime()->getMicroSeconds()->toString());
+        $this->assertSame('200997', $provider->getTime()->getMicroseconds()->toString());
 
         $provider->setUsec(30192);
 
         $this->assertSame('1050804050', $provider->getTime()->getSeconds()->toString());
-        $this->assertSame('30192', $provider->getTime()->getMicroSeconds()->toString());
+        $this->assertSame('30192', $provider->getTime()->getMicroseconds()->toString());
 
         $this->assertNotSame($time, $provider->getTime());
     }